move prod tests to e2e folder
