+++ backend/tests/e2e/prod/VercelAzureSql.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeAll, afterAll } from 'vitest';
-import db from '../../db/index.js';
+import db from '../../../db/index.js';
 import fetch from 'node-fetch';
 
 // This file tests Azure SQL connection in the Vercel environment
